requirements.txt
Django>=4.2,<5.0
Pillow
django-ckeditor>=6.7.0
django-admin-sortable2>=1.1.0